requirements.txt
# --- Backend runtime ---
fastapi>=0.111
uvicorn[standard]>=0.30
python-dotenv>=1.0
pydantic>=2.7

# --- Security / rate limiting ---
slowapi>=0.1.9

# --- HTTP clients ---
httpx>=0.27
requests>=2.32

# --- Vector DB ---
chromadb>=0.5.5

# --- Embeddings / NLP ---
transformers>=4.44
together>=1.2.0
numpy>=1.26

# --- PDF parsing ---
pymupdf>=1.24

# --- LlamaParse + LlamaIndex core (your parser + chunker) ---
llama-parse>=0.4.9
llama-index-core>=0.10.62

# --- Reflex UI (frontend) ---
reflex>=0.5
